refactor: rename RouteRegistrator with RouteRegistrant
